requirements.txt
Flask==2.3.3
slack_sdk==3.26.1
python-dotenv==1.0.1
